refactor: Refactor Swap Service and Bottom Button Sections
- Renamed `CompleteSwapBottomButtonSection`, `ReceivedSwapBottomButtonSection`, `AfterSwapBottomButtonSection`, and `BeforeSwapBottomButtonSection` to new file path.
- Changed the HTTP method of `swapAccept`, `swapReject`, and `swapComplete` from `POST` to `PATCH` in `SwapService`.
- Modified `swapComplete`, `swapAccept`, `swapReject`, `initiateChatSwapFlow` to handle null `tradesId`.
- Added `fetchProductDetail` in `ShoppingDetailScreen` with `LaunchedEffect`.
- Modified `fetchProductDetail` function to public in `ShoppingDetailViewModel` and added `init` function.
