+++ app/src/main/java/com/swapit/company/ui/shopping/detail/bottom/AfterSwapBottomButtonSection.kt
@@ -1,24 +1,24 @@
-package com.swapit.company.ui.shopping.detail
+package com.swapit.company.ui.shopping.detail.bottom
 
 import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.layout.padding
-import androidx.compose.material3.Icon
-import androidx.compose.material3.TextButton
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.LocalConfiguration
-import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
 import androidx.navigation.NavHostController
 import com.swapit.company.R
 import com.swapit.company.ui.chat.ChatViewModel
+import com.swapit.company.ui.component.DefaultButton
 import com.swapit.company.ui.component.ModalButton
+import com.swapit.company.ui.shopping.detail.ShoppingDetailViewModel
 import com.swapit.company.ui.swap.SwapViewModel
 import com.swapit.company.ui.theme.Gray5
 import com.swapit.company.ui.theme.Paddings
 
 @Composable
-fun ReceivedSwapBottomButtonSection(
+fun AfterSwapBottomButtonSection(
     viewModel: ShoppingDetailViewModel,
     swapViewModel: SwapViewModel,
     chatViewModel: ChatViewModel,
@@ -29,7 +29,7 @@ fun ReceivedSwapBottomButtonSection(
     val screenWidthDp = configuration.screenWidthDp.dp
     val horizontalPadding = screenWidthDp.value / 20
     ModalButton(
-        text = "스왑 거절하기",
+        text = "스왑 요청 취소하기",
         contentPadding =
             PaddingValues(
                 horizontal = horizontalPadding.dp,
@@ -37,34 +37,23 @@ fun ReceivedSwapBottomButtonSection(
             ),
         containerColor = Gray5,
     ) {
-        swapViewModel.swapReject(viewModel.detailContents.trade!!.tradesId)
+        swapViewModel.swapCancel(viewModel.detailContents.trade!!.tradesId)
         navController.navigateUp()
     }
-    ModalButton(
-        text = "스왑 수락하기",
-        contentPadding =
-            PaddingValues(
-                horizontal = horizontalPadding.dp,
-                vertical = Paddings.xlarge,
-            ),
-        containerColor = Gray5,
-    ) {
-        swapViewModel.swapAccept(viewModel.detailContents.trade!!.tradesId)
-    }
-    TextButton(
-        onClick = {
-            if (shoppingDetailViewModel.detailContents.trade == null) {
-                chatViewModel.initiateChatFlow(shoppingDetailViewModel.goodsId.toLong(), navController)
-            } else {
-                chatViewModel.initiateChatSwapFlow(shoppingDetailViewModel.detailContents.trade!!.tradesId, navController)
-            }
-        },
+    DefaultButton(
+        text = stringResource(R.string.shopping_detail_chat_bottom_button),
         enabled = true,
         modifier = Modifier.padding(start = Paddings.large),
+        contentPadding =
+            PaddingValues(
+                horizontal = horizontalPadding.dp * 2,
+                vertical = Paddings.xlarge,
+            ),
     ) {
-        Icon(
-            painter = painterResource(R.drawable.ic_chat),
-            contentDescription = "채팅 아이콘",
-        )
+        if (shoppingDetailViewModel.detailContents.trade == null) {
+            chatViewModel.initiateChatFlow(shoppingDetailViewModel.goodsId.toLong(), navController)
+        } else {
+            chatViewModel.initiateChatSwapFlow(shoppingDetailViewModel.detailContents.trade?.tradesId ?: 1, navController)
+        }
     }
 }
