+++ app/src/main/java/com/swapit/company/ui/shopping/detail/bottom/CompleteSwapBottomButtonSection.kt
@@ -1,4 +1,4 @@
-package com.swapit.company.ui.shopping.detail
+package com.swapit.company.ui.shopping.detail.bottom
 
 import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.layout.padding
@@ -12,21 +12,23 @@ import com.swapit.company.R
 import com.swapit.company.ui.chat.ChatViewModel
 import com.swapit.company.ui.component.DefaultButton
 import com.swapit.company.ui.component.ModalButton
-import com.swapit.company.ui.navigation.NavItem
+import com.swapit.company.ui.shopping.detail.ShoppingDetailViewModel
+import com.swapit.company.ui.swap.SwapViewModel
 import com.swapit.company.ui.theme.Gray5
 import com.swapit.company.ui.theme.Paddings
 
 @Composable
-fun BeforeSwapBottomButtonSection(
+fun CompleteSwapBottomButtonSection(
     navController: NavHostController,
     shoppingDetailViewModel: ShoppingDetailViewModel,
+    swapViewModel: SwapViewModel,
     chatViewModel: ChatViewModel,
 ) {
     val configuration = LocalConfiguration.current
     val screenWidthDp = configuration.screenWidthDp.dp
     val horizontalPadding = screenWidthDp.value / 20
     ModalButton(
-        text = stringResource(R.string.shopping_detail_swap_request_bottom_button),
+        text = "스왑 완료하기",
         contentPadding =
             PaddingValues(
                 horizontal = horizontalPadding.dp,
@@ -34,7 +36,8 @@ fun BeforeSwapBottomButtonSection(
             ),
         containerColor = Gray5,
     ) {
-        navController.navigate(NavItem.MyProductSelection.screenRoute + "/${shoppingDetailViewModel.goodsId}")
+        swapViewModel.swapComplete(shoppingDetailViewModel.detailContents.trade?.tradesId ?: 1)
+        navController.popBackStack()
     }
     DefaultButton(
         text = stringResource(R.string.shopping_detail_chat_bottom_button),
@@ -45,12 +48,11 @@ fun BeforeSwapBottomButtonSection(
                 horizontal = horizontalPadding.dp * 2,
                 vertical = Paddings.xlarge,
             ),
-        onClick = {
-            if (shoppingDetailViewModel.detailContents.trade == null) {
-                chatViewModel.initiateChatFlow(shoppingDetailViewModel.goodsId.toLong(), navController)
-            } else {
-                chatViewModel.initiateChatSwapFlow(shoppingDetailViewModel.detailContents.trade!!.tradesId, navController)
-            }
-        },
-    )
+    ) {
+        if (shoppingDetailViewModel.detailContents.trade == null) {
+            chatViewModel.initiateChatFlow(shoppingDetailViewModel.goodsId.toLong(), navController)
+        } else {
+            chatViewModel.initiateChatSwapFlow(shoppingDetailViewModel.detailContents.trade?.tradesId ?: 1, navController)
+        }
+    }
 }
