+++ app/src/main/java/com/swapit/company/ui/shopping/detail/bottom/ReceivedSwapBottomButtonSection.kt
@@ -1,23 +1,25 @@
-package com.swapit.company.ui.shopping.detail
+package com.swapit.company.ui.shopping.detail.bottom
 
 import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.layout.padding
+import androidx.compose.material3.Icon
+import androidx.compose.material3.TextButton
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.LocalConfiguration
-import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.unit.dp
 import androidx.navigation.NavHostController
 import com.swapit.company.R
 import com.swapit.company.ui.chat.ChatViewModel
-import com.swapit.company.ui.component.DefaultButton
 import com.swapit.company.ui.component.ModalButton
+import com.swapit.company.ui.shopping.detail.ShoppingDetailViewModel
 import com.swapit.company.ui.swap.SwapViewModel
 import com.swapit.company.ui.theme.Gray5
 import com.swapit.company.ui.theme.Paddings
 
 @Composable
-fun AfterSwapBottomButtonSection(
+fun ReceivedSwapBottomButtonSection(
     viewModel: ShoppingDetailViewModel,
     swapViewModel: SwapViewModel,
     chatViewModel: ChatViewModel,
@@ -28,7 +30,7 @@ fun AfterSwapBottomButtonSection(
     val screenWidthDp = configuration.screenWidthDp.dp
     val horizontalPadding = screenWidthDp.value / 20
     ModalButton(
-        text = "스왑 요청 취소하기",
+        text = "스왑 거절하기",
         contentPadding =
             PaddingValues(
                 horizontal = horizontalPadding.dp,
@@ -36,23 +38,35 @@ fun AfterSwapBottomButtonSection(
             ),
         containerColor = Gray5,
     ) {
-        swapViewModel.swapCancel(viewModel.detailContents.trade!!.tradesId)
+        swapViewModel.swapReject(viewModel.detailContents.trade?.tradesId ?: 1)
         navController.navigateUp()
     }
-    DefaultButton(
-        text = stringResource(R.string.shopping_detail_chat_bottom_button),
-        enabled = true,
-        modifier = Modifier.padding(start = Paddings.large),
+    ModalButton(
+        text = "스왑 수락하기",
         contentPadding =
             PaddingValues(
-                horizontal = horizontalPadding.dp * 2,
+                horizontal = horizontalPadding.dp,
                 vertical = Paddings.xlarge,
             ),
+        containerColor = Gray5,
     ) {
-        if (shoppingDetailViewModel.detailContents.trade == null) {
-            chatViewModel.initiateChatFlow(shoppingDetailViewModel.goodsId.toLong(), navController)
-        } else {
-            chatViewModel.initiateChatSwapFlow(shoppingDetailViewModel.detailContents.trade!!.tradesId, navController)
-        }
+        swapViewModel.swapAccept(viewModel.detailContents.trade?.tradesId ?: 1)
+        shoppingDetailViewModel.fetchProductDetail()
+    }
+    TextButton(
+        onClick = {
+            if (shoppingDetailViewModel.detailContents.trade == null) {
+                chatViewModel.initiateChatFlow(shoppingDetailViewModel.goodsId.toLong(), navController)
+            } else {
+                chatViewModel.initiateChatSwapFlow(shoppingDetailViewModel.detailContents.trade!!.tradesId, navController)
+            }
+        },
+        enabled = true,
+        modifier = Modifier.padding(start = Paddings.large),
+    ) {
+        Icon(
+            painter = painterResource(R.drawable.ic_chat),
+            contentDescription = "채팅 아이콘",
+        )
     }
 }
